requirements.txt
python-telegram-bot==22.5
fastapi==0.95.1
uvicorn[standard]==0.22.0
yt-dlp==2025.08.01
python-dotenv==1.0.0
